source/machine_connector/m2c2_publisher/requirements.txt
greengrasssdk>=1.6.0
python-dateutil==2.8.1
backoff==1.10.0